+++ lending/RELEASES.md
@@ -449,3 +449,13 @@ Based on:
 - [typescript v0.4.21] lending
 ### Releases
 - [NPM v0.4.21] https://www.npmjs.com/package/ryan-lending/v/0.4.21 - lending
+
+## 2024-10-29 01:00:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.423.1 (2.442.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.0] lending
+### Releases
+- [NPM v0.5.0] https://www.npmjs.com/package/ryan-lending/v/0.5.0 - lending